+++ test/amplitude.js
@@ -2185,45 +2185,6 @@ describe('setVersionName', function() {
       assert.lengthOf(server.requests, 2);
       var events = JSON.parse(querystring.parse(server.requests[1].requestBody).e);
       assert.deepEqual(events[0].user_properties, {});
-
-      // verify session storage set
-      assert.deepEqual(JSON.parse(sessionStorage.getItem('amplitude_utm_properties')), expectedProperties);
-    });
-
-    it('should not set utmProperties if utmProperties data already in session storage', function() {
-      reset();
-      clock.tick(30 * 60 * 1000 + 1);
-      var existingProperties = {
-        utm_campaign: 'old',
-        utm_content: 'bottom',
-        utm_medium: 'texts',
-        utm_source: 'datamonster',
-        utm_term: 'conditions'
-      };
-      sessionStorage.setItem('amplitude_utm_properties', JSON.stringify(existingProperties));
-
-      cookie.set('__utmz', '133232535.1424926227.1.1.utmcct=top&utmccn=new');
-      var utmParams = '?utm_source=amplitude&utm_medium=email&utm_term=terms';
-      amplitude.getInstance()._initUtmData(utmParams);
-
-      assert.lengthOf(server.requests, 1);
-      var events = JSON.parse(querystring.parse(server.requests[0].requestBody).e);
-      assert.lengthOf(events, 1);
-
-      // first event should be identify with initial_utm properties and NO existing utm properties
-      assert.equal(events[0].event_type, '$identify');
-      assert.deepEqual(events[0].user_properties, {
-        '$setOnce': {
-          initial_utm_campaign: 'new',
-          initial_utm_content: 'top',
-          initial_utm_medium: 'email',
-          initial_utm_source: 'amplitude',
-          initial_utm_term: 'terms'
-        }
-      });
-
-      // should not override any existing utm properties values in session storage
-      assert.equal(sessionStorage.getItem('amplitude_utm_properties'), JSON.stringify(existingProperties));
     });
   });
 
@@ -2279,69 +2240,6 @@ describe('setVersionName', function() {
       // second event should be the test event with no referrer information
       assert.equal(events[1].event_type, 'Referrer Test Event');
       assert.deepEqual(events[1].user_properties, {});
-
-      // referrer should be propagated to session storage
-      assert.equal(sessionStorage.getItem('amplitude_referrer'), JSON.stringify(expected));
-    });
-
-    it('should not set referrer if referrer data already in session storage', function() {
-      reset();
-      clock.tick(30 * 60 * 1000 + 1);
-      sessionStorage.setItem('amplitude_referrer', 'https://www.google.com/search?');
-      amplitude.init(apiKey, undefined, {includeReferrer: true, batchEvents: true, eventUploadThreshold: 2});
-      amplitude.logEvent('Referrer Test Event', {});
-      assert.lengthOf(server.requests, 1);
-      var events = JSON.parse(querystring.parse(server.requests[0].requestBody).e);
-      assert.lengthOf(events, 2);
-
-      // first event should be identify with initial_referrer and NO referrer
-      assert.equal(events[0].event_type, '$identify');
-      assert.deepEqual(events[0].user_properties, {
-        '$setOnce': {
-          'initial_referrer': 'https://amplitude.com/contact',
-          'initial_referring_domain': 'amplitude.com'
-        }
-      });
-
-      // second event should be the test event with no referrer information
-      assert.equal(events[1].event_type, 'Referrer Test Event');
-      assert.deepEqual(events[1].user_properties, {});
-    });
-
-    it('should not override any existing initial referrer values in session storage', function() {
-      reset();
-      sessionStorage.setItem('amplitude_referrer', 'https://www.google.com/search?');
-      amplitude.init(apiKey, undefined, {includeReferrer: true, batchEvents: true, eventUploadThreshold: 3});
-      amplitude.getInstance()._saveReferrer('https://facebook.com/contact');
-      amplitude.logEvent('Referrer Test Event', {});
-      assert.lengthOf(server.requests, 1);
-      var events = JSON.parse(querystring.parse(server.requests[0].requestBody).e);
-      assert.lengthOf(events, 3);
-
-      // first event should be identify with initial_referrer and NO referrer
-      assert.equal(events[0].event_type, '$identify');
-      assert.deepEqual(events[0].user_properties, {
-        '$setOnce': {
-          'initial_referrer': 'https://amplitude.com/contact',
-          'initial_referring_domain': 'amplitude.com'
-        }
-      });
-
-      // second event should be another identify but with the new referrer
-      assert.equal(events[1].event_type, '$identify');
-      assert.deepEqual(events[1].user_properties, {
-        '$setOnce': {
-          'initial_referrer': 'https://facebook.com/contact',
-          'initial_referring_domain': 'facebook.com'
-        }
-      });
-
-      // third event should be the test event with no referrer information
-      assert.equal(events[2].event_type, 'Referrer Test Event');
-      assert.deepEqual(events[2].user_properties, {});
-
-      // existing value persists
-      assert.equal(sessionStorage.getItem('amplitude_referrer'), 'https://www.google.com/search?');
     });
   });